Prefer using non-`Option` known value of negotiated version more in the handshake.
`is_tls13()` operates on the negotiated version that is stored as an `Option` and
treats `None` as equivalent to TLS 1.2. Avoid using it within the handshake state
machine so that we don't have to reason about what to do when it is `None`.

Similarly, avoid `sess.common.negotiated_version.unwrap()` in favor of the
known-not-`None` value.

In the near future I hope to eliminate the remaining uses of `is_tls13()` outside
of the handshake state machine.